Updated API Keys UI
The UI has now been moved to the settings page.
