+++ src/router.js
@@ -18,8 +18,7 @@ import NotFound from "./pages/NotFound.vue";
 import DockerHosts from "./components/settings/Docker.vue";
 import MaintenanceDetails from "./pages/MaintenanceDetails.vue";
 import ManageMaintenance from "./pages/ManageMaintenance.vue";
-import ManageAPIKeys from "./pages/ManageAPIKeys.vue";
-import AddAPIKey from "./pages/AddAPIKey.vue";
+import APIKeys from "./components/settings/APIKeys.vue";
 import Plugins from "./components/settings/Plugins.vue";
 
 // Settings - Sub Pages
@@ -116,6 +115,10 @@ const routes = [
                                 component: Security,
                             },
                             {
+                                path: "api-keys",
+                                component: APIKeys,
+                            },
+                            {
                                 path: "proxies",
                                 component: Proxies,
                             },
@@ -156,14 +159,6 @@ const routes = [
                     {
                         path: "/maintenance/edit/:id",
                         component: EditMaintenance,
-                    },
-                    {
-                        path: "/apikeys",
-                        component: ManageAPIKeys
-                    },
-                    {
-                        path: "/apikeys/add",
-                        component: AddAPIKey
                     },
                 ],
             },
